pyproject.toml
[tool.poetry]
name = "demo-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.80.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}
SQLAlchemy = "^1.4.40"
aiomysql = "^0.1.1"
SQLAlchemy-Utils = "^0.38.3"
python-socketio = "^5.7.1"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.19.0"
aiosqlite = "^0.17.0"
httpx = "^0.23.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 